+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/IntegerSchemaSpec.scala
@@ -4,6 +4,7 @@ package schemas
 import org.json4s.JsonDSL._
 import org.json4s._
 
+import PropertySets._
 import UnitSpec._
 
 class IntegerSchemaSpec extends UnitSpec {