+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/StringSchemaSpec.scala
@@ -5,6 +5,7 @@ import org.json4s.{DefaultFormats, Formats}
 import org.json4s.JsonDSL._
 import org.json4s._
 
+import PropertySets._
 import UnitSpec._
 
 class StringSchemaSpec extends UnitSpec {